LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/viewport-units-dynamic-expected.txt



FAIL CSS Container Queries Test: @container-dependent elements respond to viewport unit changes assert_equals: vw after width resize expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
